pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbxctl"
version = "0.1.0"
description = "dbxctl - Idempotent management of users, groups, and memberships in Databricks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Guilherme Mancuso", email = "guilherme.mancuso@gmail.com" },
]
dependencies = [
    "databricks-sdk>=0.20.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
dbxctl = "dbxctl:main"

[tool.hatch.build.targets.wheel]
packages = ["dbx_iam"]
extra-includes = ["dbxctl.py"]
